+++ qiskit/circuit/__init__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2017, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""Module for Circuits."""
+from .quantumcircuit import QuantumCircuit
+from .classicalregister import ClassicalRegister
+from .quantumregister import QuantumRegister
+from .gate import Gate
+from .instruction import Instruction
+from .instructionset import InstructionSet
+from .measure import Measure
+from .reset import Reset
+from .compositegate import CompositeGate
