+++ qiskit/circuit/classicalregister.py
@@ -10,7 +10,7 @@ Classical register reference object.
 """
 import itertools
 
-from ._register import Register
+from .register import Register
 
 
 class ClassicalRegister(Register):
