+++ qiskit/circuit/gate.py
@@ -8,8 +8,8 @@
 """
 Unitary gate.
 """
-from ._instruction import Instruction
-from ._qiskiterror import QiskitError
+from qiskit._qiskiterror import QiskitError
+from .instruction import Instruction
 
 
 class Gate(Instruction):
