+++ qiskit/circuit/instruction.py
@@ -29,9 +29,9 @@ Instructions are identified by the following fields, and are serialized as such 
 import sympy
 
 from qiskit.qasm._node import _node
-from ._qiskiterror import QiskitError
-from ._quantumregister import QuantumRegister
-from ._classicalregister import ClassicalRegister
+from qiskit._qiskiterror import QiskitError
+from .quantumregister import QuantumRegister
+from .classicalregister import ClassicalRegister
 
 
 class Instruction(object):
